+++ changes.md
@@ -3,5 +3,6 @@ Rikulo Changes
 
 **July 20, 2012**
 
+* Rename Motion/Scroller's onMoving to onMove and moving to move.
 * `ProfileDeclaration.anchor` is default to `parent` if `location` is specified.
 * `Activity.onMount_` was removed. Please override `Activity.onCreate_` instead.